+++ src/pages/index.tsx
@@ -17,3 +17,4 @@ export { default as DetailedAnalysis } from "./DetailedAnalysis/DetailedAnalysis
 export { default as StudentRegister } from "./StudentRegister/StudentRegister";
 export { default as PasswordReset } from "./PasswordReset/PasswordReset";
 export { default as Profile } from "./Profile/";
+export { default as CreateNewBatch } from "./Batches/CreateBatches";
